+++ RELEASES.md
@@ -1,9 +1,10 @@
 # Release Notes
 
-## Release 0.3.2 (2021-11-??)
+## Release 0.3.2 (2022-01-01)
 
 - Added comparison between atan::atan and cordic::atan.
 - Changed vector argument of fft to `&mut [Complex<T>]`.
+- Changed descrtiption of the `niirf` error plot.
 
 ## Release 0.3.1 (2021-11-29)